requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1.2
environs==8.0.0
idna==2.10
marshmallow==3.8.0
psycopg2-binary==2.8.6
pyTelegramBotAPI==3.7.3
python-dotenv==0.14.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.10
django-phonenumber-field==5.0.0
phonenumbers==8.12.11
Pillow==7.2.0